+++ Sources/App/Models/APIResponse.swift
@@ -1,8 +1,6 @@
 import Foundation
 import Vapor
 
-typealias FutureAPIResponse<T: Codable> = Future<APIResponse<T>>
-
 /// The response object that should be returned to the client.
 ///
 /// Formats the response json as:
@@ -13,14 +11,4 @@ typealias FutureAPIResponse<T: Codable> = Future<APIResponse<T>>
 /// ```
 struct APIResponse<T: Codable>: Content {
     let data: T
-}
-
-extension Future where T: Codable {
-    
-    func toAPIResponse() -> Future<APIResponse<T>> {
-        self.map(to: APIResponse<T>.self) { object in
-            return APIResponse<T>(data: object)
-        }
-    }
-    
 }
